Login Page: Change way of handling clicking on buttons to remove the redirects
This commit compltely changes what happens when you click on the Register or Login button, it now uses an ajax request to call the dotnet code and doesn't redirect automatically after the request, which means it is gonna be much better to inform the user about the outcome.
The data is passed using the body of the ajax requests, maybe we could move the models of these to a separate file.