Login Page: Inform the user about the outcome of the actions
This commit introuduces a modal box that will show when an error occured, bad password, or any other error. It will also open instructions on verifying the account (we need to implement that at some point) when a registration is successful. The only case where it doesn't show the modal box is when a login operation was successful, it will jst reload the page in that case but i think it is fine as the user will see that he is connected in the top right corner