AssetManager (BIG WIP): Backend / Frontend data transfer work + models
This commit implements some of the required things about AssetManager and what's related to it, we have models for all the different type of assets, using a base class that will hopefully make refactoring easier.
These assets classes will be used throughout lot of code since these assets will then be used by other parts of the website that wants to use them.
This commit also includes some new methods related to the assets in DbSelector and DbInserter.
Right now the AssetManager service doesn't cache anything but it is planned to be changed at some point to avoid extra database calls.
On the frontend side, there isn't much new aside from the game and character selection; these are currently wired to getting the data for the specific game / character but there is nothing drawn yet.
The way the Assets are serialized and deserialized for json objects is i think a bit hackish, it's using a custom JsonConverter to get around the lazy type properties but you know what, it does work.
On the database and FileManager side, there's 3 new tables and 4 new folders to store the data about the assets, the files might need to be stored in their actual extension, we will see how HTML handles that

This description is already long enough and there are things changed that i didn't write about, maybe this should've been splitted in several commits but oh well... here we are now :p