AssetManager: Render assets + add FileExtension to the assets + small fixes
Note: This currently only renders the site assets on screen but the code will be extracted to be used for the games and character assets

This adds the file extension to the assets because they are needed by HTML to properly render the images and this actually goes through and render the images as well as their name in the div, this uses a lot of css magic to make it work so I expect someone that actually know what they're doing to fix this.
Next step will be to implement the actual functionality aka adding, deleting and editing the assets.
This also fixes the AssetManager service not using the constants in the FileManager